Changed critter genetic traits to only inherit if dominant gene is present.  Added some comments and defaults.  Added logic for elemental flags.  Changed crit and dodge to integers, and made 0 the unable value.

+++ db/migrate/20110903025323_create_critters.rb
@@ -5,8 +5,8 @@ class CreateCritters < ActiveRecord::Migration
       t.string :url
       t.integer :level
       t.integer :hp
-      t.decimal :dodge
-      t.decimal :crit
+      t.integer :dodge
+      t.integer :crit
       t.integer :physical_damage
       t.integer :fire_damage
       t.integer :earth_damage
